requirements.txt
Flask==3.0.2
python-dotenv==1.0.1

Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.4
Flask-Migrate==4.0.7

Flask-Login==0.6.3
Werkzeug==3.0.1

marshmallow==3.13.0
flask-marshmallow==1.2.0
marshmallow-sqlalchemy==1.0.0
gunicorn==21.2.0
PyJWT>=2.8.0
Flask-Cors
Flask-JWT-Extended